refactor(db): Split migration.rs into directory module
